requirements.txt
alembic==1.7.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
Mako==1.1.5
MarkupSafe==2.0.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.19.0
requests==2.25.1
SQLAlchemy==1.4.12
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
